Source code cleaning

+++ rat_utilities/sources/rat_math_utilities.c
@@ -299,9 +299,9 @@ uint8_t rat_calculate_crc (uint16_t value,
 
     for (counter_bit = 8;counter_bit > 0;--counter_bit) {
       if (checksum & 0x80) {
-        checksum = ((uint8_t)checksum << 1) ^ polynomial;
+        checksum = (checksum << 1) ^ polynomial;
       } else {
-        checksum = ((uint8_t)checksum << 1);
+        checksum = (checksum << 1);
       }
     }
   }
